fix: (multiply plugin) move type to plugin folder
